fix(backend): semver version sort was not stable for ios
- for ios, sort the versions using a stable sorting algorithm

fixes #3516
